+++ example/dargon2_example.dart
@@ -1,14 +1,22 @@
+// Copyright 2020 Tejas Mehta <tmthecoder@gmail.com>
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
 import 'dart:convert';
 
 import 'package:dargon2/dargon2.dart';
 
-void main() {}
+void main() {
+  hashString();
+  hashWithBytes();
+}
 
 void hashString() {
   var password = 'password';
   //use Salt(List<int> bytes) for a salt from an Integer list
   var s = Salt.newSalt();
-  //Hash with pre-set params (iterations: 32, memory: 1024, parallelism: 2, length: 32, type: Argon2Type.i, version: Argon2Version.V13)
+  //Hash with pre-set params (iterations: 32, memory: 256, parallelism: 2, length: 32, type: Argon2Type.i, version: Argon2Version.V13)
   var result = argon2.hashPasswordString(password, salt: s); //type: DArgon2Result
 
   //Raw hash values available as int list, base 64 string, and hex string
@@ -25,7 +33,7 @@ void hashString() {
   var stringEncoded = result.encodedString; // type: String
   print('Encoded UTF-8 String: $stringEncoded');
 
-  //Verify password (returns true/false)
+  //Verify password (returns true/false), uses default type (Argon2Type.i)
   var verified = argon2.verifyHashString(password, stringEncoded); // type: bool
   print('password is verified: $verified');
 }
@@ -59,5 +67,6 @@ void hashWithBytes() {
   print('Encoded UTF-8 String: $stringEncoded');
 
   //Verify password (returns true/false)
-  var verified = argon2.verifyHashBytes(password, bytesEncoded); // type: bool
+  var verified = argon2.verifyHashBytes(password, bytesEncoded, type: Argon2Type.id); // type: bool
+  print('password is verified: $verified');
 }