+++ lib/dargon2.dart
@@ -1,8 +1,19 @@
-/// Support for doing something awesome.
+// Copyright 2020 Tejas Mehta <tmthecoder@gmail.com>
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+/// Library for Argon2 Hashing and verification in dart prrojects
 ///
-/// More dartdocs go here.
+/// To use this library in your code:
+///
+///     import 'package:dargon2/dartgon2.dart';
+///
+/// For detailed usage please look at the example
 library argon2_dart;
 
 export 'src/argon2.dart';
 export 'src/utils/salt.dart';
 export 'src/utils/dargon2_result.dart';
+export 'src/exceptions/DArgon2ErrorCodes.dart';
+export 'src/exceptions/DArgon2Exception.dart';